+++ examples/stocks/lib/main.dart
@@ -7,7 +7,6 @@ library stocks;
 import 'dart:math' as math;
 import 'dart:sky' as sky;
 
-import 'package:sky/editing/input.dart';
 import 'package:sky/theme/colors.dart' as colors;
 import 'package:sky/theme/typography.dart' as typography;
 import 'package:sky/widgets.dart';
