+++ sky/packages/sky/lib/widgets.dart
@@ -22,6 +22,7 @@ export 'package:sky/src/widgets/drawer.dart';
 export 'package:sky/src/widgets/drawer_divider.dart';
 export 'package:sky/src/widgets/drawer_header.dart';
 export 'package:sky/src/widgets/drawer_item.dart';
+export 'package:sky/src/widgets/editable_text.dart';
 export 'package:sky/src/widgets/flat_button.dart';
 export 'package:sky/src/widgets/floating_action_button.dart';
 export 'package:sky/src/widgets/focus.dart';
@@ -31,6 +32,7 @@ export 'package:sky/src/widgets/homogeneous_viewport.dart';
 export 'package:sky/src/widgets/icon.dart';
 export 'package:sky/src/widgets/icon_button.dart';
 export 'package:sky/src/widgets/ink_well.dart';
+export 'package:sky/src/widgets/input.dart';
 export 'package:sky/src/widgets/material.dart';
 export 'package:sky/src/widgets/material_button.dart';
 export 'package:sky/src/widgets/mimic.dart';
